Just using int type to simplify further usage

+++ src/Components/UnitType.h
@@ -1,5 +1,5 @@
 #pragma once
 
 namespace Sample::Components {
-	using UnitType = int16_t;
+	using UnitType = int;
 }
